# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Maicare server API documentation.",
        "title": "Maicare API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your-email@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/attachments/upload": {
            "post": {
                "description": "Upload a file to the server",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_UploadHandlerResponse"
                        }
                    }
                }
            }
        },
        "/attachments/{id}": {
            "get": {
                "description": "Get an attachment by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Get an attachment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attachment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_GetAttachmentByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an attachment by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attachments"
                ],
                "summary": "Delete an attachment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Attachment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_DeleteAttachmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.DeleteAttachmentResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_id": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "api.GetAttachmentByIdResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_id": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_DeleteAttachmentResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.DeleteAttachmentResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_GetAttachmentByIdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.GetAttachmentByIdResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_UploadHandlerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.UploadHandlerResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.UploadHandlerResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_id": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Add 'Bearer ' prefix before your JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}